+++ core/src/transaction_priority.rs
@@ -1,0 +1,47 @@
+use {
+    solana_cost_model::cost_model::CostModel,
+    solana_runtime::bank::Bank,
+    solana_runtime_transaction::{
+        transaction_meta::TransactionConfiguration, transaction_with_meta::TransactionWithMeta,
+    },
+};
+
+/// Calculate priority and cost for a transaction:
+///
+/// Cost is calculated through the `CostModel`,
+/// and priority is calculated through a formula here that attempts to sell
+/// blockspace to the highest bidder.
+///
+/// The priority is calculated as:
+/// P = R / (1 + C)
+/// where P is the priority, R is the reward,
+/// and C is the cost towards block-limits.
+///
+/// Current minimum costs are on the order of several hundred,
+/// so the denominator is effectively C, and the +1 is simply
+/// to avoid any division by zero due to a bug - these costs
+/// are calculated by the cost-model and are not direct
+/// from user input. They should never be zero.
+/// Any difference in the prioritization is negligible for
+/// the current transaction costs.
+pub(crate) fn calculate_priority_and_cost(
+    bank: &Bank,
+    transaction: &impl TransactionWithMeta,
+    transaction_configuration: &TransactionConfiguration,
+) -> (u64, u64) {
+    let cost = CostModel::calculate_cost(transaction, &bank.feature_set).sum();
+    let reward = bank.calculate_reward_for_transaction(transaction, transaction_configuration);
+
+    // We need a multiplier here to avoid rounding down too aggressively.
+    // For many transactions, the cost will be greater than the fees in terms of raw lamports.
+    // For the purposes of calculating prioritization, we multiply the fees by a large number so that
+    // the cost is a small fraction.
+    // An offset of 1 is used in the denominator to explicitly avoid division by zero.
+    const MULTIPLIER: u64 = 1_000_000;
+    (
+        reward
+            .saturating_mul(MULTIPLIER)
+            .saturating_div(cost.saturating_add(1)),
+        cost,
+    )
+}
